requirements.txt
fastapi
uvicorn
pydantic
pydantic-settings
pillow
transformers
torch
torchvision
python-multipart
accelerate
bitsandbytes